ai_agent/requirements.txt
requests
python-dotenv
pytest
pytest-cov
pytest-mock
pytest-timeout
pytest-asyncio